# docs/02-design/openapi-campusswap.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CampusSwap 文档管理平台 API",
    "version": "1.0.0-m3",
    "description": "系统与权限域（M3 已实现，25 条）。字段口径与 docs/02-design/API_SPECIFICATION.md 逐字一致；所有 ID 在 JSON 中为字符串，时间为 yyyy-MM-dd HH:mm:ss。文档域 30 条随 M4 实现追加。"
  },
  "servers": [
    { "url": "http://localhost:10087", "description": "本机 dev（Spring Boot 4.1.1，端口 10087）" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["用户管理"],
        "summary": "用户列表",
        "description": "权限点 sys:user。默认按 updatedAt 倒序；部门名与角色集合批量 IN 补齐（SQL 预算 4~5 条常数）。",
        "parameters": [
          { "name": "pageNum", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "keyword", "in": "query", "schema": { "type": "string", "maxLength": 64 }, "description": "匹配 username 或 realName" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ACTIVE", "LOCKED", "DISABLED"] } },
          { "name": "deptId", "in": "query", "schema": { "type": "string", "pattern": "^\\d{1,19}$" } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageVoUserVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" }
        }
      },
      "post": {
        "tags": ["用户管理"],
        "summary": "新增用户",
        "description": "权限点 sys:user:add。密码 BCrypt(strength 10) 落库；roles 是角色【编码】数组，不传默认绑定 STAFF。",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreateDtoReq" },
            "example": { "username": "u1002", "realName": "李娜", "deptId": "2", "email": "lina@example.com", "phone": "13800000002", "roles": ["STAFF"], "password": "Staff@123" } } }
        },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultUserVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["用户管理"],
        "summary": "用户详情",
        "description": "权限点 sys:user。响应不含 passwordHash。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultUserVo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["用户管理"],
        "summary": "编辑用户",
        "description": "权限点 sys:user:edit。username 与密码不可经本接口修改；部门或角色变化后即时清该用户权限缓存。roles 省略=不改角色，空数组=清空。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdateDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultUserVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/users/{id}/status": {
      "put": {
        "tags": ["用户管理"],
        "summary": "停用/启用用户",
        "description": "权限点 sys:user:disable。非 ACTIVE 时提交后强制下线（清全部 token）并清权限缓存；不允许改自己的状态（400）。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserStatusDtoReq" }, "example": { "status": "DISABLED" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultUserVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/users/{id}/password": {
      "put": {
        "tags": ["用户管理"],
        "summary": "重置密码",
        "description": "权限点 sys:user:reset。重置成功后该用户全部 token 失效，须用新密码重新登录。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserPasswordDtoReq" } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/roles": {
      "get": {
        "tags": ["角色管理"],
        "summary": "角色列表",
        "description": "权限点 sys:role。排序：内置角色在前，再按 sortOrder 升序。列表不返回权限规模。",
        "parameters": [
          { "name": "pageNum", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "keyword", "in": "query", "schema": { "type": "string", "maxLength": 64 } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageVoRoleVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" }
        }
      },
      "post": {
        "tags": ["角色管理"],
        "summary": "新增角色",
        "description": "权限点 sys:role:add。新建角色 isBuiltin=0、初始无权限。",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RoleDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultRoleVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/roles/{id}": {
      "put": {
        "tags": ["角色管理"],
        "summary": "编辑角色",
        "description": "权限点 sys:role:edit。code 必须与当前一致（编码不可改）；内置角色不可改名。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RoleDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultRoleVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["角色管理"],
        "summary": "删除角色",
        "description": "权限点 sys:role:delete。内置角色或被用户/部门引用时 409。软删除时同步改写 code 释放唯一键。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/roles/{id}/permissions": {
      "get": {
        "tags": ["角色管理"],
        "summary": "查询角色权限",
        "description": "权限点 sys:role。返回该角色在 sys_role_permission 中的全部权限 ID（含父节点与按钮节点）。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultRolePermissionVo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["角色管理"],
        "summary": "角色授权（覆盖式）",
        "description": "权限点 sys:role:grant。事务内先删后插；保存成功后清空该角色下全部用户的权限缓存 → 授权即时生效，无需重新登录。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RolePermissionDtoReq" }, "example": { "permissionIds": ["1", "10", "100"] } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "登录返回的 token，格式：Authorization: Bearer <token>" }
    },
    "parameters": {
      "PathId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^\\d+$" }, "description": "字符串形式的数字 ID" }
    },
    "responses": {
      "OkVoid": { "description": "成功（data 为 null）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" } } } },
      "BadRequest": { "description": "参数校验失败 / 业务规则拒绝（message 为中文原文）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 400, "message": "页码不能超过100，请缩小筛选范围后再试", "data": null } } } },
      "Unauthorized": { "description": "未登录或 token 失效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 401, "message": "登录状态已失效，请重新登录", "data": null } } } },
      "NoPermission": { "description": "已登录但缺权限点", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 403, "message": "无权限执行该操作", "data": null } } } },
...
      "NotFound": { "description": "资源不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 404, "message": "用户不存在或已被删除", "data": null } } } },
      "Conflict": { "description": "唯一约束或状态冲突", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 409, "message": "登录名已存在，请更换", "data": null } } } }
    },
    "schemas": {
      "ResponseResultVoid": { "type": "object", "properties": { "code": { "type": "integer", "example": 200 }, "message": { "type": "string", "example": "成功" }, "data": { "nullable": true } } },
...
      "ResponseResultUserVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/UserVo" } } },
      "ResponseResultRoleVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/RoleVo" } } },
...
      "ResponseResultRolePermissionVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/RolePermissionVo" } } },
...
      "PageVoUserVo": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" }, "message": { "type": "string" },
          "data": { "type": "object", "properties": { "list": { "type": "array", "items": { "$ref": "#/components/schemas/UserVo" } }, "total": { "type": "integer" }, "pageNum": { "type": "integer" }, "pageSize": { "type": "integer" } } }
        }
      },
      "PageVoRoleVo": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" }, "message": { "type": "string" },
          "data": { "type": "object", "properties": { "list": { "type": "array", "items": { "$ref": "#/components/schemas/RoleVo" } }, "total": { "type": "integer" }, "pageNum": { "type": "integer" }, "pageSize": { "type": "integer" } } }
        }
      },
      "AuditVo": {
        "type": "object",
        "description": "实体型 VO 的公共审计字段（时间与 ID 均为字符串）",
        "properties": {
          "createdAt": { "type": "string", "example": "2026-09-21 14:20:05" },
          "createdBy": { "type": "string", "example": "1" },
          "updatedAt": { "type": "string", "example": "2026-09-21 14:20:05" },
          "updatedBy": { "type": "string", "example": "1" }
        }
      },
...
      "UserVo": {
        "allOf": [{ "$ref": "#/components/schemas/AuditVo" }],
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "username": { "type": "string" }, "realName": { "type": "string" },
          "deptId": { "type": "string" }, "deptName": { "type": "string" },
          "email": { "type": "string", "nullable": true }, "phone": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "LOCKED", "DISABLED"] },
          "roles": { "type": "array", "items": { "type": "string" } },
          "lastLoginAt": { "type": "string", "nullable": true }
        }
      },
      "RoleVo": {
        "allOf": [{ "$ref": "#/components/schemas/AuditVo" }],
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "name": { "type": "string" }, "code": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "isBuiltin": { "type": "integer", "enum": [0, 1] },
          "sortOrder": { "type": "integer" }
        }
      },
      "RolePermissionVo": {
        "type": "object",
        "properties": { "roleId": { "type": "string" }, "permissionIds": { "type": "array", "items": { "type": "string" } } }
      },
...
      "UserCreateDtoReq": {
        "type": "object", "required": ["username", "realName", "deptId", "password"],
        "properties": {
          "username": { "type": "string", "pattern": "^[A-Za-z0-9_]{4,64}$", "description": "登录名不能为空，且只能包含字母、数字与下划线" },
          "realName": { "type": "string", "maxLength": 64, "description": "姓名不能为空且不超过64个字符" },
          "deptId": { "type": "string", "description": "部门不存在，请重新选择" },
          "roles": { "type": "array", "items": { "type": "string" }, "description": "角色【编码】数组，如 [\"STAFF\"]；不传 = 仅绑定 STAFF" },
          "phone": { "type": "string", "pattern": "^$|^1\\d{10}$", "description": "手机号格式不正确" },
          "email": { "type": "string", "maxLength": 128, "description": "邮箱格式不正确" },
          "password": { "type": "string", "pattern": "^(?=.*[A-Za-z])(?=.*\\d)\\S{8,64}$", "description": "初始密码至少8位且需同时包含字母和数字" }
        }
      },
      "UserUpdateDtoReq": {
        "type": "object", "required": ["realName", "deptId"],
        "properties": {
          "realName": { "type": "string", "maxLength": 64, "description": "姓名不能为空且不超过64个字符" },
          "deptId": { "type": "string", "description": "部门不存在，请重新选择" },
          "roles": { "type": "array", "items": { "type": "string" }, "description": "角色【编码】数组；省略=不改角色，空数组=清空" },
          "phone": { "type": "string", "pattern": "^$|^1\\d{10}$" },
          "email": { "type": "string", "maxLength": 128 }
        }
      },
      "UserStatusDtoReq": {
        "type": "object", "required": ["status"],
        "properties": { "status": { "type": "string", "enum": ["ACTIVE", "LOCKED", "DISABLED"], "description": "账号状态取值非法，仅支持 ACTIVE、LOCKED、DISABLED" } }
      },
      "UserPasswordDtoReq": {
        "type": "object", "required": ["newPassword"],
        "properties": { "newPassword": { "type": "string", "pattern": "^(?=.*[A-Za-z])(?=.*\\d)\\S{8,64}$", "description": "新密码至少8位且需同时包含字母和数字" } }
      },
      "RoleDtoReq": {
        "type": "object", "required": ["name", "code"],
        "properties": {
          "name": { "type": "string", "maxLength": 64, "description": "角色名称不能为空且不超过64个字符" },
          "code": { "type": "string", "pattern": "^[A-Z0-9_]{2,32}$", "description": "角色编码只能包含大写字母、数字与下划线；创建后不可修改" },
          "description": { "type": "string", "maxLength": 255, "description": "角色描述不能超过255个字符" }
        }
      },
      "RolePermissionDtoReq": {
        "type": "object", "required": ["permissionIds"],
        "properties": { "permissionIds": { "type": "array", "items": { "type": "string" }, "description": "权限【ID】数组；空数组 = 清空该角色全部权限" } }
      },
...
    }
  }
}